+++ gems/kettle-jem/CHANGELOG.md
@@ -22,6 +22,14 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Changed
 
+- [kc] kettle-jem/prepare: updated 9 project files:
+  - dependencies (9)
+
+- [kc] kettle-jem/template: updated 3 project files:
+  - code and tests (1)
+  - dependencies (1)
+  - other (1)
+
 ### Deprecated
 
 ### Removed