+++ gems/kettle-jem/CHANGELOG.md
@@ -22,8 +22,8 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Changed
 
-- [kc] kettle-jem/prepare: updated 9 project files:
-  - dependencies (9)
+- [kc] kettle-jem/prepare: updated 13 project files:
+  - dependencies (13)
 
 - [kc] kettle-jem/template: updated 3 project files:
   - code and tests (1)